AI Layer/Prescription Extraction Service/requirements.txt
python-dotenv
requests
google-generativeai
python-socketio
eventlet
